Initial Commit with Gemfile and file structure
This way of rebuilding sinatra is some sort of API driven development.
So app.rb is a normal sinatra app file and it determines how much
should be implemented in lib/sinatra.rb or even files under lib
directories.

Gemfile is imported cause that we need rake and puma to actually run
a sinatra app.

Within this commit,

 - A minimum functional Sinatra module and Base class are defined.
 - Rack handler, Puma, is used to back up our sample app.